+++ src/OSharp.CodeGeneration/Templates/AdminController.cshtml
@@ -36,7 +36,7 @@ using OSharp.Core.Functions;
 using OSharp.Core.Modules;
 using OSharp.Data;
 using OSharp.Filter;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 using @(entity.Module.Namespace);
 using @(entity.Module.Namespace).Dtos;
