+++ src/OSharp.Permissions/Security/DataAuthCacheBase.cs
@@ -21,7 +21,7 @@ using OSharp.Entity;
 using OSharp.Extensions;
 using OSharp.Filter;
 using OSharp.Identity;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace OSharp.Security
