+++ src/OSharp.Permissions/Security/EntityRoleBase.cs
@@ -14,7 +14,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using OSharp.Entity;
 using OSharp.Extensions;
 using OSharp.Filter;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace OSharp.Security
