+++ src/OSharp.Permissions/Security/Events/DataAuthCacheRefreshEventData.cs
@@ -10,7 +10,7 @@
 using System.Collections.Generic;
 
 using OSharp.EventBuses;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace OSharp.Security.Events
