+++ src/OSharp.Permissions/Security/Events/DataAuthCacheRefreshEventHandler.cs
@@ -11,7 +11,7 @@ using System;
 
 using OSharp.Dependency;
 using OSharp.EventBuses;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace OSharp.Security.Events
