+++ src/OSharp.Permissions/Security/Events/FunctionAuthCacheRefreshEventHandler.cs
@@ -14,7 +14,7 @@ using Microsoft.Extensions.DependencyInjection;
 using OSharp.AspNetCore;
 using OSharp.Dependency;
 using OSharp.EventBuses;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace OSharp.Security.Events
