+++ src/OSharp.Permissions/Security/IEntityRoleStore.cs
@@ -14,7 +14,7 @@ using System.Threading.Tasks;
 
 using OSharp.Data;
 using OSharp.Filter;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace OSharp.Security
