+++ src/OSharp/Extensions/StringExtensions.cs
@@ -19,7 +19,7 @@ using System.Web;
 using Newtonsoft.Json;
 
 using OSharp.Collections;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace OSharp.Extensions
