+++ tests/OSharp.Tests/Secutiry/AesHelperTests.cs
@@ -10,7 +10,7 @@ using Shouldly;
 using Xunit;
 
 
-namespace OSharp.Secutiry.Tests
+namespace OSharp.Security.Tests
 {
     public class AesHelperTests
     {
