+++ tests/OSharp.Tests/Secutiry/TransmissionCryptorTests.cs
@@ -5,7 +5,7 @@ using Shouldly;
 using Xunit;
 
 
-namespace OSharp.Secutiry.Tests
+namespace OSharp.Security.Tests
 {
     public class TransmissionCryptorTests
     {
