+++ tests/web/Liuliu.Demo.Core/Security/SecurityPack.cs
@@ -15,7 +15,7 @@ using Liuliu.Demo.Security.Entities;
 using OSharp.Core.EntityInfos;
 using OSharp.Core.Functions;
 using OSharp.Security;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace Liuliu.Demo.Security
