+++ tests/web/Liuliu.Demo.Web/Areas/Admin/Controllers/Security/RoleEntityController.cs
@@ -29,7 +29,7 @@ using OSharp.Data;
 using OSharp.Entity;
 using OSharp.Extensions;
 using OSharp.Filter;
-using OSharp.Secutiry;
+using OSharp.Security;
 
 
 namespace Liuliu.Demo.Web.Areas.Admin.Controllers
